Updated scripts and reorganized files
