+++ Pandator/Assets/NetGun/SE/Scripts/NetGunHits.cs
@@ -1,6 +1,19 @@
 using UnityEngine;
 
-public class NetGunHits : MonoBehaviour
+public class NetGunHits:MonoBehaviour
 {
 	[SerializeField] private AudioSource audioSource;
+	private bool isHit = false;
+	private void Update()
+	{
+		if (Input.GetKeyUp(KeyCode.H))
+		{
+			isHit = true;
+		}
+		if (isHit)
+		{
+			audioSource.Play();
+			isHit = false;
+		}
+	}
 }
