--- Pandator/Assets/SE/Scripts/NetGunHits.cs
@@ -1,19 +1,0 @@
-using UnityEngine;
-
-public class NetGunHits:MonoBehaviour
-{
-	[SerializeField] private AudioSource audioSource;
-	private bool isHit = false;
-	private void Update()
-	{
-		if (Input.GetKeyUp(KeyCode.H))
-		{
-			isHit = true;
-		}
-		if (isHit)
-		{
-			audioSource.Play();
-			isHit = false;
-		}
-	}
-}
