--- Pandator/Assets/SE/Scripts/NetGunHits.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a57a0755b29dd8d448473ed28831ac83
